review: manage date computed
