+++ frontend/app/components/home/ImportantInformationSection/ITNCard.vue
@@ -1,13 +1,9 @@
 <script setup lang="ts">
 import Card from "../Card.vue";
 
-const {
-    yesterday,
-    yesterdayTemperature,
-    gap,
-    yesterdayLastYear,
-    temperatureChangeYearOverYear,
-} = useHomeData();
+const { yesterdayTemperature, gap, temperatureChangeYearOverYear } =
+    useHomeData();
+const { yesterday, yesterdayLastYear } = useCustomDate();
 </script>
 <template>
     <Card
